MEDIUM / Visible support for DrawingTreeNode

+++ diana-swing/src/main/java/org/openflexo/fge/swing/view/JDrawingView.java
@@ -715,13 +715,15 @@ public class JDrawingView<M> extends JDianaLayeredView<M> implements Autoscroll,
 				}
 			});
 			for (GeometricNode<?> gn : geomList) {
-				// TODO: use the same graphics, just change DrawingTreeNode
-				// JFGEGeometricGraphics geometricGraphics = new JFGEGeometricGraphics(gn, this);
-				JFGEGeometricGraphics geometricGraphics = getGeometricGraphics(gn);
-				geometricGraphics.createGraphics(g2/*, controller*/);
-				gn.paint(geometricGraphics);
-				geometricGraphics.releaseGraphics();
-				geometricGraphics.delete();
+				if (gn.shouldBeDisplayed()) {
+					// TODO: use the same graphics, just change DrawingTreeNode
+					// JFGEGeometricGraphics geometricGraphics = new JFGEGeometricGraphics(gn, this);
+					JFGEGeometricGraphics geometricGraphics = getGeometricGraphics(gn);
+					geometricGraphics.createGraphics(g2/*, controller*/);
+					gn.paint(geometricGraphics);
+					geometricGraphics.releaseGraphics();
+					geometricGraphics.delete();
+				}
 			}
 		}
 	}
